Modified the IndexViewModel
It is now comprised of 2 collections of SubjectViewModels, each of which is comprised of the subject name, and a collection of exams in that subject.

I built it all while attempting to implement SRP, and the TDD approach (this is why so many files are commited all at once).

Student Main Page now displays a Future Exams table, and a Past Exams table.